Change kuka_kss_rsi_driver to kuka_rsi_driver
