+++ kuka_rsi_driver/hardware_interface.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
-<library path="kuka_kss_rsi_driver">
-  <class type="kuka_kss_rsi_driver::KukaRSIHardwareInterface"
+<library path="kuka_rsi_driver">
+  <class type="kuka_rsi_driver::KukaRSIHardwareInterface"
          base_class_type="hardware_interface::SystemInterface">
     <description>ros2_control hardware interface for KSS-powered robots</description>
   </class>
