+++ kuka_rsi_driver/krl/KSS/ros_rsi.src
@@ -45,7 +45,7 @@ DEF RSI_AxisCorr( )
 ;
 ; ROS INDUSTRIAL: KUKA RSI HW INTERFACE
 ; Realtime UDP data exchange with
-; kuka_kss_rsi_driver
+; kuka_rsi_driver
 ;
 ; =============================================
