+++ kuka_rsi_driver/test/config/test2_joint_trajectory_controller_config.yaml
