+++ sonar-project.properties
@@ -4,15 +4,15 @@ sonar.projectKey=kroshu_kuka_drivers
 
 sonar.host.url=https://sonarcloud.io
 
-sonar.modules=examples,kuka_sunrise_fri_driver, kuka_kss_rsi_driver, kuka_iiqka_eac_driver, kuka_drivers_core, controllers
+sonar.modules=examples,kuka_sunrise_fri_driver, kuka_rsi_driver, kuka_iiqka_eac_driver, kuka_drivers_core, controllers
 
 examples.sonar.projectName=examples
 examples.sonar.sources=./iiqka_moveit_example/src, ./iiqka_moveit_example/launch
 kuka_sunrise_fri_driver.sonar.projectName=kuka_sunrise_fri_driver
 kuka_sunrise_fri_driver.sonar.sources=./src,./include,./launch
 kuka_sunrise_fri_driver.sonar.exclusions=**/fri_client_sdk/**
-kuka_kss_rsi_driver.sonar.projectName=kuka_kss_rsi_driver
-kuka_kss_rsi_driver.sonar.sources=./src,./include
+kuka_rsi_driver.sonar.projectName=kuka_rsi_driver
+kuka_rsi_driver.sonar.sources=./src,./include
 kuka_iiqka_eac_driver.sonar.projectName=kuka_iiqka_eac_driver
 kuka_iiqka_eac_driver.sonar.sources=./src,./include/kuka_iiqka_eac_driver
 kuka_iiqka_eac_driver.sonar.exclusions=**/mock/**
